rename thunk to fetch
